@@ -2,43 +2,46 @@
   "name": "my-app",
   "version": "0.0.1",
   "private": true,
+  "type": "module",
   "scripts": {
     "start": "react-native start",
-    "dev": "react-native start",
+    "dev": "watchman watch-del \"$(pwd)\"; watchman watch-project \"$(pwd)\"; clear; react-native start --reset-cache",
     "android": "react-native run-android",
     "ios": "react-native run-ios",
-    "prettify": "prettier --write src/ test/",
+    "prettify": "prettier -w '*.{js,cjs,ts}' src/",
+    "check": "tsc --noEmit",
+    "lint": "eslint .",
     "spec": "jest",
-    "lint": "eslint src/ test/",
-    "prettycheck": "prettier --check src/ test/",
-    "typecheck": "tsc --noEmit",
-    "test": "npm run lint && npm run prettycheck && npm run typecheck && npm run spec"
+    "test": "npm run check && npm run lint && npm run spec"
   },
   "dependencies": {
-    "react": "^18.3.1",
-    "react-native": "0.76.1"
+    "react": "18.3.1",
+    "react-native": "0.76.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.26.0",
-    "@babel/preset-env": "^7.26.0",
-    "@babel/runtime": "^7.26.0",
+    "@babel/core": "^7.25.2",
+    "@babel/preset-env": "^7.25.3",
+    "@babel/runtime": "^7.25.0",
     "@jest/globals": "^29.7.0",
-    "@mgcrea/eslint-config-react-native": "^0.12.16",
-    "@react-native/metro-config": "^0.76.1",
+    "@mgcrea/eslint-config-react-native": "^0.12.20",
+    "@react-native-community/cli": "15.0.1",
+    "@react-native-community/cli-platform-android": "15.0.1",
+    "@react-native-community/cli-platform-ios": "15.0.1",
+    "@react-native/babel-preset": "0.76.3",
+    "@react-native/metro-config": "0.76.3",
+    "@react-native/typescript-config": "0.76.3",
     "@testing-library/jest-native": "^5.4.3",
-    "@testing-library/react-native": "^12.8.1",
-    "@tsconfig/react-native": "^3.0.5",
-    "@types/jest": "^29.5.14",
-    "@types/react": "^18.3.12",
-    "@types/react-test-renderer": "^18.3.0",
+    "@testing-library/react-native": "^12.9.0",
+    "@types/react": "^18.2.6",
+    "@types/react-test-renderer": "^18.0.0",
     "babel-jest": "^29.7.0",
-    "babel-plugin-tsconfig-paths": "^1.0.3",
-    "eslint": "^9.14.0",
+    "eslint": "^9.16.0",
     "jest": "^29.7.0",
-    "metro-react-native-babel-preset": "^0.77.0",
-    "prettier": "^3.3.3",
+    "prettier": "3.4.2",
     "prettier-plugin-organize-imports": "^4.1.0",
-    "react-test-renderer": "^18.3.1",
-    "typescript": "^5.6.3"
+    "typescript": "^5.7.2"
+  },
+  "engines": {
+    "node": ">=18"
   }
 }
