@@ -1,5 +1,10 @@
-/** @type {import('@babel/core').TransformOptions} */
-module.exports = {
-  presets: ["module:metro-react-native-babel-preset"],
-  plugins: ["tsconfig-paths"],
+// @ts-check
+
+/**
+ * Babel configuration
+ * @see https://babeljs.io/docs/options
+ * @type {import('@babel/core').TransformOptions}
+ */
+export default {
+  presets: ["module:@react-native/babel-preset"],
 };
