@@ -1,5 +1,5 @@
 import { AppRegistry } from "react-native";
-import { name as appName } from "./app.json";
+import { name as appName } from "./app.json" assert { type: "json" };
 import { App } from "./src/App";
 
-AppRegistry.registerComponent(appName, () => App);
+AppRegistry.registerComponent(String(appName), () => App);
