@@ -1,10 +1,10 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
+// @ts-check
+/* eslint-disable no-undef, @typescript-eslint/no-require-imports  */
 const { getDefaultConfig, mergeConfig } = require("@react-native/metro-config");
 
 /**
  * Metro configuration
- * https://facebook.github.io/metro/docs/configuration
- *
+ * @see https://reactnative.dev/docs/metro
  * @type {import('metro-config').MetroConfig}
  */
 const config = {};
